0.4
Finished the homepage. Made it fully responsive and nice looking on all platforms. Fixed bugs where there was overflow to the right. Added a README.md including a short description of the project, the license, my Github link under authors, acknowledgements for the font and the icons used, and then my logo.